Removed NotImplementedError from calculation of GAST.
Though the error stated that complementary terms in the equation of the equinox
were necessary to implement GAST using the IAU 2000B nutation model, they are *not*
included in the SOFA implementation iauGst00b. Documentation in greenwich_sidereal_time
updated to indicate exactly which method is being used to calculate GAST,
sofatest.py updated to use apparent=True instead of apparent='simple'.